refactor commandDir, make dotenv loading predictable
