+++ packages/cli/package.json
@@ -47,6 +47,7 @@
     "envinfo": "7.8.1",
     "execa": "5.1.1",
     "fast-glob": "3.2.11",
+    "findup-sync": "5.0.0",
     "fs-extra": "10.1.0",
     "humanize-string": "2.1.0",
     "latest-version": "5.1.0",
@@ -63,7 +64,8 @@
     "secure-random-password": "0.2.3",
     "terminal-link": "2.1.1",
     "title-case": "3.0.3",
-    "yargs": "17.5.1"
+    "yargs": "17.5.1",
+    "yargs-parser": "21.1.1"
   },
   "devDependencies": {
     "@babel/cli": "7.18.10",
