+++ packages/cli/src/commands/setup/ui/ui.js
@@ -1,10 +1,16 @@
 import terminalLink from 'terminal-link'
 
 export const command = 'ui <library>'
+
 export const description = 'Set up a UI design or style library'
-export const builder = (yargs) =>
+
+export async function builder(yargs) {
+  const setupUIChakraUICommand = await import('./libraries/chakra-ui')
+  const setupUIMantineCommand = await import('./libraries/mantine')
+  const setupUITailwindCSSCommand = await import('./libraries/tailwindcss')
+  const setupUIWindiCSSCommand = await import('.//libraries/windicss')
+
   yargs
-    .commandDir('./libraries')
     .demandCommand()
     .epilogue(
       `Also see the ${terminalLink(
@@ -12,3 +18,8 @@ export const builder = (yargs) =>
         'https://redwoodjs.com/docs/cli-commands#setup-ui'
       )}`
     )
+    .command(setupUIChakraUICommand)
+    .command(setupUIMantineCommand)
+    .command(setupUITailwindCSSCommand)
+    .command(setupUIWindiCSSCommand)
+}
